Merged Latest Luigi to Trunk
Fixed: Issue 58. Issue 95. Issue 133.

Added Co-op check box to game page, is returned in API and is available as parameter in Advanced Search.

Added Column to indicate banner availability on both Search and Browse Platforms.

Added Ability to sort any listed games (search, advanced search, browse by platforms). Have not made all sort options available that were requested in the issue as it is not good practise to let the user sort by options that can't be viewed in the results.